Sampling time dt now can be given from the command line; long command line options; improved help

+++ src/b2t.c
@@ -5,19 +5,19 @@ int main(int argc, char **argv) {
 	
 	int i, j=0, data=5 ; 
 	FILE *datain ;  
-	double *cand ; 
+	float *cand ; 
 	
-    cand 		= (double *) calloc (data, sizeof(double));
+    cand 		= (float *) calloc (data, sizeof(float));
     	
     // reading from the input datafile   
 	if ((datain = fopen (argv[1], "rb")) != NULL) {
 
 		while (!feof(datain)) {
 
-		   if((fread ((void *)(cand), sizeof (double), data, datain))==data)
+		   if((fread ((void *)(cand), sizeof (float), data, datain))==data)
 			
 				for(i=0; i<data; i++) 
-					printf("%le ", cand[i]) ;
+					printf("%.8le ", cand[i]) ;
 					
 					printf("\n") ;
